+++ src/components/project/rlgym-learn/subconfigs/BaseConfigEditor.tsx
@@ -1,8 +1,7 @@
-import { type BaseConfigModel } from "../../../../models/rlgym-learn/api";
-import ToggleField from "../../../config-cards/ToggleCard";
 import StringField from "../../../config-cards/StringField";
 import NumberField from "../../../config-cards/NumberField";
 import NumberCard from "../../../config-cards/NumberCard";
+import type { BaseConfigModel } from "rlgym-learn-client";
 
 export interface BaseConfigEditorArgs{
     baseConfig: BaseConfigModel
@@ -22,8 +21,6 @@ function BaseConfigEditor({baseConfig, setBaseConfig}:BaseConfigEditorArgs) {
 
                 <p className="fw-lighter mt-2">Miscellaneous</p>
                 <div className="gray-bg rounded">
-                    <ToggleField value={baseConfig.batched_tensor_action_associated_learning_data} help="No idea what that does?" icon="collection" onToggle={() => setBaseConfig({...baseConfig, batched_tensor_action_associated_learning_data: !baseConfig.batched_tensor_action_associated_learning_data})} text="Batch action associated learning data"></ToggleField>
-                    <hr className="mx-2 my-1"></hr>
                     <StringField value={baseConfig.flinks_folder} help="Folder where the file links are stored" text="File links folder" icon="folder" onChange={(value) => {if(value === null) return; else setBaseConfig({...baseConfig, flinks_folder: value})}} required></StringField>
                     <hr className="mx-2 my-1"></hr>
                     <NumberField value={baseConfig.random_seed} required help="The seed used for the RNG" icon="dice-1" onChange={(value) => {if(value === null) return; else setBaseConfig({...baseConfig, random_seed: value})}} text="Random seed"></NumberField>